update function to controllers
