+++ services/contacts.json
@@ -60,8 +60,8 @@
     "phone": "(748) 206-2688"
   },
   {
-    "id": "21",
-    "name": "Victor Selischev",
+    "id": "20",
+    "name": "Alex Gusarov",
     "email": "vvs.kharkov@gmail.com",
     "phone": "(067) 895-37-44"
   }
